rework generators
