--- lib/generators/draper/decorator/templates/decorator_spec.rb
@@ -1,4 +1,0 @@
-require 'spec_helper'
-
-describe <%= resource_name.singularize.camelize %>Decorator do
-end
