+++ lib/generators/draper/templates/application_decorator.rb
