+++ lib/generators/draper/templates/application_decorator_spec.rb
