+++ lib/generators/draper/templates/application_decorator_test.rb
