+++ lib/generators/rspec/templates/decorator_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe <%= singular_name.camelize %>Decorator do
+end
