+++ lib/generators/test_unit/templates/decorator_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class <%= resource_name.singularize.camelize %>DecoratorTest < ActiveSupport::TestCase
+class <%= singular_name.camelize %>DecoratorTest < ActiveSupport::TestCase
   def setup
     ApplicationController.new.set_current_view_context
   end
