+++ spec/generators/draper/decorator/decorator_generator_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
 # Generators are not automatically loaded by Rails
-require 'generators/draper/decorator/decorator_generator'
+require 'generators/draper/decorator_generator'
 
-describe Draper::DecoratorGenerator do
+describe Draper::Generators::DecoratorGenerator do
   # Tell the generator where to put its output (what it thinks of as Rails.root)
   destination File.expand_path("../../../../../tmp", __FILE__)
 
@@ -21,17 +21,7 @@ describe Draper::DecoratorGenerator do
   end
 
   context 'decorator name' do
-    before { run_generator ["YourModel"]  }
-
-    describe 'spec/decorators/your_model_decorator_spec.rb' do
-      subject { file('spec/decorators/your_model_decorator_spec.rb') }
-      it { should exist }
-      it { should contain "describe YourModelDecorator" }
-    end
-  end
-
-  context 'default test framework' do
-    before { run_generator ["YourModel"]  }
+    before { run_generator ["YourModel", '-t=rspec']  }
 
     describe 'spec/decorators/your_model_decorator_spec.rb' do
       subject { file('spec/decorators/your_model_decorator_spec.rb') }
@@ -60,54 +50,3 @@ describe Draper::DecoratorGenerator do
     end
   end
 end
-
-
-=begin
-  describe 'no arguments' do
-    before { run_generator %w(products)  }
-
-    describe 'app/decorators/products_decorator.rb' do
-      subject { file('app/decorators/products_decorator.rb') }
-      it { should exist }
-      it { should contain "class ProductsDecorator < ApplicationDecorator" }
-    end
-  end
-
-
-  context 'simple' do
-    before { run_generator %w(products)  }
-
-    describe 'app/decorators/products_decorator.rb' do
-      subject { file('app/decorators/products_decorator.rb') }
-      it { should exist }
-      it { should contain "class ProductsDecorator < ApplicationDecorator" }
-    end
-  end
-
-
-
-
-
-  context 'using rspec' do
-
-    describe 'app/decorators/products_decorator.rb' do
-      subject { file('app/decorators/products_decorator.rb') }
-      it { should exist }
-      it { should contain "class ProductsDecorator < ApplicationDecorator" }
-    end
-
-    shared_examples_for "ApplicationDecoratorGenerator" do
-      describe 'app/decorators/application_decorator.rb' do
-        subject { file('app/decorators/application_decorator.rb') }
-        it { should exist }
-        it { should contain "class ApplicationDecorator < Draper::Base" }
-      end
-    end
-
-    describe 'spec/decorators/application_decorator_spec.rb' do
-      subject { file('spec/decorators/application_decorator_spec.rb') }
-      it { should exist }
-      it { should contain "describe ApplicationDecorator do" }
-    end
-  end
-=end
